// src/shared/enums/region.enum.ts
export enum Region {
    VINNYTSIA = "Vinnytsia",
    VOLYN = "Volyn",
    DNIPROPETROVSK = "Dnipropetrovsk",
    DONETSK = "Donetsk",
    ZHYTOMYR = "Zhytomyr",
    ZAKARPATTIA = "Zakarpattia",
    ZAPORIZHZHIA = "Zaporizhzhia",
    IVANO_FRANKIVSK = "Ivano-Frankivsk",
    KYIV = "Kyiv",
    KIROVOHRAD = "Kirovohrad",
    LUHANSK = "Luhansk",
    LVIV = "Lviv",
    MYKOLAIV = "Mykolaiv",
    ODESA = "Odesa",
    POLTAVA = "Poltava",
    RIVNE = "Rivne",
    SUMY = "Sumy",
    TERNOPIL = "Ternopil",
    KHMELNYTSKYI = "Khmelnytskyi",
    CHERKASY = "Cherkasy",
    CHERNIVTSI = "Chernivtsi",
    CHERNIHIV = "Chernihiv",
    KRYM = "Crimea",
    SEVASTOPOL = "Sevastopol"
}